requirements.txt
tables
configparser
setuptools
numpy
pandas
scipy
matplotlib
sklearn
pyyaml